mastodon-message/requirements.txt
nonebot2[fastapi,httpx,websockets]>=2.4.4,<3.0.0
nonebot-adapter-onebot>=2.4.6,<3.0.0
websockets>=15.0,<16.0
